backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

pdfplumber==0.10.3
scikit-learn==1.3.2
PyPDF2==3.0.1

nltk==3.8.1
python-Levenshtein==0.21.1
pydantic==2.5.0
python-dotenv==1.0.0
